LieAlgebras: Move gap iso stuff into one file per direction
blub


blub
