+++ experimental/LieAlgebras/src/LieAlgebras.jl
@@ -73,6 +73,8 @@ include("LieAlgebra.jl")
 include("AbstractLieAlgebra.jl")
 include("LinearLieAlgebra.jl")
 include("LieAlgebraModule.jl")
+include("iso_oscar_gap.jl")
+include("iso_gap_oscar.jl")
 include("GapWrapper.jl")
 
 end
