+++ experimental/LieAlgebras/test/runtests.jl
@@ -1,4 +1,5 @@
 include("Combinatorics-test.jl")
-include("GapWrapper-test.jl")
+include("iso_oscar_gap-test.jl")
+include("iso_gap_oscar-test.jl")
 include("LieAlgebra-test.jl")
 include("LieAlgebraModule-test.jl")
